Extract category glyph rendering
