+++ service-worker.js
@@ -66,6 +66,7 @@ const APP_SHELL = [
   '/js/dashboard-widget-controls.js',
   '/js/dashboard-widget-renderers.js',
   '/js/chart-card-recs.js',
+  '/js/category-glyphs.js',
   '/js/marker-detail-modal.js',
   '/js/light-conditions-now.js',
   '/js/light-sessions-view.js',
